scripts/content-discovery/requirements.txt
websockets==12.0
aiohttp==3.9.1
pyyaml==6.0.1
requests==2.31.0
prometheus-client==0.19.0
